feat: Add Group functionality to Engine
Add GroupContext struct for grouping routes under a common prefix.
Add methods for CORS middleware, adding middleware, defining routes, and
handling Websocket connections.
