+++ go.mod
@@ -1,8 +1,13 @@
 module github.com/d3code/fireball
 
-go 1.22.2
+go 1.22.4
 
 require (
 	github.com/google/uuid v1.6.0
 	golang.org/x/net v0.25.0
+)
+
+require (
+	github.com/d3code/xlog v0.0.0-20240616235105-9390f807bd37 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
